// components/ui/types.ts
export type ColorProps = {
  name: string;
  colors: {
    light: {
      text: string;
      background: string;
      primary: string;
      secondary: string;
      accent: string;
    };
    dark: {
      text: string;
      background: string;
      primary: string;
      secondary: string;
      accent: string;
    };
  };
};